+++ models/business-unit/src/division/builder.spec.ts
@@ -34,6 +34,7 @@ describe('builder', () => {
         topLevelUnit: expect.objectContaining({
           typeId: 'business-unit',
         }),
+        approvalRuleMode: 'Explicit',
         custom: null,
         createdAt: expect.any(String),
         createdBy: expect.objectContaining({