+++ models/category/src/index.ts
@@ -1,5 +1,6 @@
 export * as CategoryDraft from './category-draft';
 export * as Category from '.';
+export * as CategorySearch from './category-search';
 
 export { default as random } from './builder';
 export { default as presets } from './presets';