+++ core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @commercetools-test-data/core
 
+## 10.5.2
+
+## 10.5.1
+
 ## 10.5.0
 
 ## 10.4.0